Support Wiimote accelerometer reporting

+++ cios-lib/types.h
@@ -36,7 +36,7 @@
 /* Data types */
 typedef char			s8;
 typedef short			s16;
-typedef long			s32;
+typedef int			s32;
 typedef unsigned char		u8;
 typedef unsigned short		u16;
 typedef unsigned int		u32;
